requirements.txt
pillow
TkinterMessagebox
windnd
sv_ttk